strips out a lot of cli code into phases to condense workload
